+++ packages/flame_tiled/lib/src/renderable_layers/renderable_layer.dart
@@ -7,30 +7,52 @@ import 'package:flame_tiled/src/renderable_layers/group_layer.dart';
 import 'package:flame_tiled/src/renderable_layers/image_layer.dart';
 import 'package:flame_tiled/src/renderable_layers/object_layer.dart';
 import 'package:flame_tiled/src/renderable_layers/tile_layers/tile_layer.dart';
+import 'package:flame_tiled/src/renderable_layers/tile_layers/unsupported_layer.dart';
 import 'package:flame_tiled/src/tile_animation.dart';
 import 'package:flame_tiled/src/tile_atlas.dart';
-import 'package:meta/meta.dart';
 import 'package:tiled/tiled.dart';
 
-@internal
-abstract class RenderableLayer<T extends Layer> {
+abstract class RenderableLayer<T extends Layer> extends PositionComponent
+    with HasPaint<RenderableLayer<T>> {
   final T layer;
   final Vector2 destTileSize;
   final TiledMap map;
 
-  /// The parent [Group] layer (if it exists)
-  final GroupLayer? parent;
+  /// A [camera] is only required if using parallax effects.
+  /// If one exists in the game tree, [camera] will be assigned to it [onLoad].
+  CameraComponent? camera;
 
   /// The [FilterQuality] that should be used by all the layers.
   final FilterQuality filterQuality;
 
+  /// The total canvas translation used in parallax effects.
+  /// This field needs to be read in the case of repeated textures
+  /// as an optimization step.
+  Vector2 totalParallax = Vector2.zero();
+
+  /// The total offsets computed from our [Layer.offsetX] and [Layer.offsetY]
+  /// and all parent layer offsets, if any.
+  Vector2 totalOffset = Vector2.zero();
+
+  /// Given the layer's [opacity], compute [Paint] for drawing.
+  /// This is useful if your layer requires translucency or other effects.
+  Paint Function(double opacity) layerPaintFactory;
+
+  /// A read on [paint] will compute the value from [layerPaintFactory].
+  @override
+  Paint get paint => layerPaintFactory(opacity);
+
   RenderableLayer({
     required this.layer,
-    required this.parent,
+    required Component? parent,
     required this.map,
+    required this.camera,
     required this.destTileSize,
+    required this.layerPaintFactory,
     FilterQuality? filterQuality,
-  }) : filterQuality = filterQuality ?? FilterQuality.none;
+  }) : filterQuality = filterQuality ?? FilterQuality.none {
+    this.parent = parent;
+  }
 
   /// [load] is a factory method to create [RenderableLayer] by type of [layer].
   static Future<RenderableLayer> load({
@@ -50,6 +72,7 @@ abstract class RenderableLayer<T extends Layer> {
       return FlameTileLayer.load(
         layer: layer,
         parent: parent,
+        camera: camera,
         map: map,
         destTileSize: destTileSize,
         animationFrames: animationFrames,
@@ -66,106 +89,123 @@ abstract class RenderableLayer<T extends Layer> {
         map: map,
         destTileSize: destTileSize,
         filterQuality: filterQuality,
+        layerPaintFactory: layerPaintFactory,
         images: images,
       );
     } else if (layer is ObjectGroup) {
       return ObjectLayer.load(
         layer,
+        parent,
+        camera,
         map,
         destTileSize,
+        layerPaintFactory,
         filterQuality,
       );
     } else if (layer is Group) {
-      final groupLayer = layer;
       return GroupLayer(
-        layer: groupLayer,
+        layer: layer,
         parent: parent,
+        camera: camera,
         map: map,
         destTileSize: destTileSize,
         filterQuality: filterQuality,
+        layerPaintFactory: layerPaintFactory,
       );
     }
 
     return UnsupportedLayer(
       layer: layer,
       parent: parent,
+      camera: camera,
       map: map,
       destTileSize: destTileSize,
+      layerPaintFactory: layerPaintFactory,
     );
   }
 
   bool get visible => layer.visible;
 
-  void render(Canvas canvas, CameraComponent? camera);
-
-  void handleResize(Vector2 canvasSize);
-
   void refreshCache();
 
-  void update(double dt);
+  @override
+  double get opacity =>
+      layer.opacity *
+      switch (parent) {
+        final GroupLayer p => p.opacity,
+        _ => 1,
+      };
 
   double get scaleX => destTileSize.x / map.tileWidth;
   double get scaleY => destTileSize.y / map.tileHeight;
 
-  late double offsetX = layer.offsetX * scaleX + (parent?.offsetX ?? 0);
-
-  late double offsetY = layer.offsetY * scaleY + (parent?.offsetY ?? 0);
-
-  late double opacity = layer.opacity * (parent?.opacity ?? 1);
-
-  late double parallaxX = layer.parallaxX * (parent?.parallaxX ?? 1);
-
-  late double parallaxY = layer.parallaxY * (parent?.parallaxY ?? 1);
+  late double offsetX = layer.offsetX * scaleX;
+  late double offsetY = layer.offsetY * scaleY;
 
   /// Calculates the offset we need to apply to the canvas to compensate for
   /// parallax positioning and scroll for the layer and the current camera
   /// position.
   /// https://doc.mapeditor.org/en/latest/manual/layers/#parallax-scrolling-factor
-  void applyParallaxOffset(Canvas canvas, CameraComponent camera) {
-    final anchor = camera.viewfinder.anchor;
-    final cameraX = camera.viewfinder.position.x;
-    final cameraY = camera.viewfinder.position.y;
-    final viewportCenterX = camera.viewport.size.x * anchor.x;
-    final viewportCenterY = camera.viewport.size.y * anchor.y;
+  void applyParallaxOffset(Canvas canvas) {
+    final viewfinder = camera?.viewfinder;
+    final cameraX = viewfinder?.position.x ?? 0;
+    final cameraY = viewfinder?.position.y ?? 0;
+    final zoom = viewfinder?.zoom ?? 1.0;
 
-    // Due to how Tiled treats the center of the view as the reference
-    // point for parallax positioning (see Tiled docs), we need to offset the
-    // entire layer
-    var x = (1 - parallaxX) * viewportCenterX;
-    var y = (1 - parallaxY) * viewportCenterY;
-    // Compensate the offset for zoom.
-    x /= camera.viewfinder.zoom;
-    y /= camera.viewfinder.zoom;
-    // Scale to tile space.
-    x /= destTileSize.x;
-    y /= destTileSize.y;
+    // Discover parent layer terms used for the calculations below.
+    final parentOffset = switch (parent) {
+      final GroupLayer p => p.totalOffset,
+      _ => Vector2.zero(),
+    };
+    final parentParallax = switch (parent) {
+      final GroupLayer p => p.totalParallax,
+      _ => Vector2.zero(),
+    };
+    final parallaxLocality = Vector2(
+      layer.parallaxX * parentParallax.x,
+      layer.parallaxY * parentParallax.y,
+    );
 
-    // Now add the scroll for the current camera position
-    x += cameraX - (cameraX * parallaxX);
-    y += cameraY - (cameraY * parallaxY);
+    // Calculate our local parallax.
+    double calcParallax(double cam, double parallax) => cam * parallax;
+    final localParallax =
+        Vector2(
+          calcParallax(cameraX, layer.parallaxX),
+          calcParallax(cameraY, layer.parallaxY),
+        ) /
+        zoom;
 
-    canvas.translate(x, y);
+    // Adjustment term for parallax locality.
+    final delta = switch (parent is GroupLayer) {
+      true => (localParallax + parallaxLocality) - parentParallax,
+      false => Vector2.zero(),
+    };
+
+    totalParallax = localParallax + parallaxLocality;
+    totalOffset = parentOffset + Vector2(offsetX, offsetY);
+
+    // Strictly apply local translations in our render step.
+    //
+    // Explanation:
+    // B/c sum of all local translations w.r.t. parallax locality is equal to the
+    // products of all absolute parallax values followed by translation, the
+    // scene graph render by Flame produces the same visuals as painting layers
+    // using absolute values in a traditional composite renderer.
+    canvas.translate(
+      (cameraX + offsetX) - (localParallax.x + delta.x),
+      (cameraY + offsetY) - (localParallax.y + delta.y),
+    );
   }
-}
 
-@internal
-class UnsupportedLayer extends RenderableLayer {
-  UnsupportedLayer({
-    required super.layer,
-    required super.parent,
-    required super.map,
-    required super.destTileSize,
-  });
-
+  // Only render if this layer is [visible].
   @override
-  void render(Canvas canvas, CameraComponent? camera) {}
-
-  @override
-  void handleResize(Vector2 canvasSize) {}
-
-  @override
-  void refreshCache() {}
-
-  @override
-  void update(double dt) {}
+  void renderTree(Canvas c) {
+    if (!visible) {
+      return;
+    }
+    c.save();
+    applyParallaxOffset(c);
+    super.renderTree(c);
+    c.restore();
+  }
 }