+++ packages/flame_tiled/lib/src/tiled_component.dart
@@ -1,6 +1,5 @@
 import 'dart:ui';
 
-import 'package:collection/collection.dart';
 import 'package:flame/cache.dart';
 import 'package:flame/components.dart';
 import 'package:flame/game.dart';
@@ -19,7 +18,7 @@ import 'package:tiled/tiled.dart';
 /// {@endtemplate}
 class TiledComponent<T extends FlameGame> extends PositionComponent
     with HasGameReference<T> {
-  /// Map instance of this component.
+  /// Map component instance which manages and draws layers.
   RenderableTiledMap tileMap;
 
   /// This property **cannot** be reassigned at runtime. To make the
@@ -66,27 +65,11 @@ class TiledComponent<T extends FlameGame> extends PositionComponent
        );
 
   @override
-  Future<void>? onLoad() async {
-    super.onLoad();
-    // Automatically use the first attached CameraComponent camera if it's not
-    // already set..
-    tileMap.camera ??= game.children.query<CameraComponent>().firstOrNull;
-  }
+  Future<void> onLoad() async {
+    await super.onLoad();
 
-  @override
-  void update(double dt) {
-    tileMap.update(dt);
-  }
-
-  @override
-  void render(Canvas canvas) {
-    tileMap.render(canvas);
-  }
-
-  @override
-  void onGameResize(Vector2 size) {
-    super.onGameResize(size);
-    tileMap.handleResize(size);
+    // Add our renderable tile map
+    await add(tileMap);
   }
 
   /// Loads a [TiledComponent] from a file.
@@ -113,6 +96,7 @@ class TiledComponent<T extends FlameGame> extends PositionComponent
     int? priority,
     bool? ignoreFlip,
     AssetBundle? bundle,
+    CameraComponent? camera,
     Images? images,
     bool Function(Tileset)? tsxPackingFilter,
     bool useAtlas = true,
@@ -130,6 +114,7 @@ class TiledComponent<T extends FlameGame> extends PositionComponent
         ignoreFlip: ignoreFlip,
         prefix: prefix,
         bundle: bundle,
+        camera: camera,
         images: images,
         tsxPackingFilter: tsxPackingFilter,
         useAtlas: useAtlas,